feat: optimize balance warning worker
