+++ src/AeFinder.BackgroundWorker.Core/Options/ScheduledTaskOptions.cs
@@ -9,9 +9,10 @@ public class ScheduledTaskOptions
     public int AppPodResourceSyncTaskPeriodMilliSeconds { get; set; } = 180000;
     public int MonthlyAutomaticChargeDay { get; set; } = 2;
     public int MonthlyAutomaticChargeTaskPeriodMilliSeconds { get; set; } = 86400000;
-    public int ChargeWarningTaskPeriodMilliSeconds { get; set; } = 86400000;
+    public int BalanceWarningTaskPeriodMilliSeconds { get; set; } = 86400000;
+    public int CleanExpiredAssetTaskPeriodMilliSeconds { get; set; } = 86400000;
     public int RenewalAdvanceWarningDays { get; set; } = 5;
-    public int BillingIndexerPollingTaskPeriodMilliSeconds { get; set; } = 10000;
+    public int OrganizationWalletSyncTaskPeriodMilliSeconds { get; set; } = 10000;
     public int OrderPaymentResultPollingTaskPeriodMilliSeconds { get; set; } = 12000;
     public int AdvanceBillPaymentResultPollingTaskPeriodMilliSeconds { get; set; } = 300000;
     public int SettlementBillPaymentResultPollingTaskPeriodMilliSeconds { get; set; } = 180000;
